+++ next_action/__init__.py
@@ -1,4 +1,4 @@
 """ Main Next-action package. """
 
 __title__ = "next-action"
-__version__ = "0.3.0"
+__version__ = "0.4.0"